LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/animation-events-expected.txt


FAIL View timeline generates and resolves finish promises and events assert_equals: A finish event is generated after end expected 1 but got 0
